# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Kino API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth/directors": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Получение всех режиссеров",
                "responses": {
                    "200": {
                        "description": "Список всех режиссеров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Director"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/film": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Получение всех фильмов",
                "responses": {
                    "200": {
                        "description": "Все фильмы успешно получены",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.FilmFull"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Обновление фильма",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фильма",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Фото фильма",
                        "name": "film_photo",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Список актеров",
                        "name": "cast_list",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID киностудии",
                        "name": "film_studio_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Продолжительность фильма в минутах",
                        "name": "duration_in_min",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ID режиссеров",
                        "name": "director_ids",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ID операторов",
                        "name": "operator_ids",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ID жанров",
                        "name": "genre_ids",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/entities.ID"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "404": {
                        "description": "Фильм не найден",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Создание фильма",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Оно\"",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Сука клоун\"",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Фото фильма",
                        "name": "film_photo",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": "\"Райан Рейнольдс, Морена Баккарин\"",
                        "description": "Список актеров",
                        "name": "cast_list",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID киностудии",
                        "name": "film_studio_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Продолжительность фильма в минутах",
                        "name": "duration_in_min",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ID режиссеров",
                        "name": "director_ids",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ID операторов",
                        "name": "operator_ids",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ID жанров",
                        "name": "genre_ids",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм успешно создан",
                        "schema": {
                            "$ref": "#/definitions/entities.ID"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для запроса",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/film-studios": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Получение всех киностудий",
                "responses": {
                    "200": {
                        "description": "Список всех киностудий",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.FilmStudio"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/film/id/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Получение фильма по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID фильма",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм по ID успешно получен",
                        "schema": {
                            "$ref": "#/definitions/entities.FilmFull"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/film/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Удаление фильма",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фильма",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фильм успешно удален",
                        "schema": {
                            "$ref": "#/definitions/entities.ID"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "404": {
                        "description": "Фильм не найден",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/genres": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Получение всех жанров",
                "responses": {
                    "200": {
                        "description": "Список всех жанров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Genre"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
        "/auth/operators": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Получение всех операторов",
                "responses": {
                    "200": {
                        "description": "Список всех операторов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Operator"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/entities.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Director": {
            "type": "object",
            "properties": {
                "fio": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "entities.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "entities.FilmFull": {
            "type": "object",
            "properties": {
                "cast_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "directors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "duration_in_min": {
                    "type": "integer"
                },
                "film_studio_name": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "photo": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}